services/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1

# Service-specific dependencies
-r investment-service/requirements.txt
-r user-service/requirements.txt
-r transaction-service/requirements.txt
-r kyc-service/requirements.txt
-r admin-service/requirements.txt
-r notification-service/requirements.txt

# Development dependencies
pytest>=6.2.4
pytest-cov>=2.12.1
httpx>=0.18.2  # For FastAPI testing 